К сожалению не удалось добится работы ORM

+++ data/src/main/java/ru/aleksandrorlov/data/database/YandexPictureDataBaseImpl.java
@@ -1,5 +1,7 @@
 package ru.aleksandrorlov.data.database;
 
+import android.util.Log;
+
 import com.orm.SugarRecord;
 
 import java.util.Collection;
@@ -7,13 +9,17 @@ import java.util.List;
 
 import javax.inject.Inject;
 
+import io.reactivex.Observable;
 import ru.aleksandrorlov.data.entity.YandexPictureEntity;
+import ru.aleksandrorlov.data.exeption.NetworkConnectionException;
+import ru.aleksandrorlov.data.net.FakeDataNet;
 
 /**
  * Created by alex on 16.03.18.
  */
 
 public class YandexPictureDataBaseImpl implements YandexPictureDataBase {
+    private final String TAG = this.getClass().getSimpleName();
 
     @Inject
     public YandexPictureDataBaseImpl() {
@@ -39,10 +45,20 @@ public class YandexPictureDataBaseImpl implements YandexPictureDataBase {
     }
 
     @Override
-    public void delete(long yandexPictureId) {
-        YandexPictureEntity yandexPictureEntity =
-                YandexPictureEntity.findById(YandexPictureEntity.class, yandexPictureId);
-        yandexPictureEntity.delete();
+    public Observable<Boolean> delete(long yandexPictureId) {
+        Log.d(TAG, "delete: contains - " + contains(yandexPictureId));
+        return Observable.create(emitter -> {
+            if (contains(yandexPictureId)){
+                YandexPictureEntity yandexPictureEntity =
+                        YandexPictureEntity.findById(YandexPictureEntity.class, yandexPictureId);
+                yandexPictureEntity.delete();
+                emitter.onNext(true);
+                emitter.onComplete();
+            } else {
+                emitter.onNext(false);
+                emitter.onComplete();
+            }
+        });
     }
 
     @Override
@@ -52,6 +68,7 @@ public class YandexPictureDataBaseImpl implements YandexPictureDataBase {
 
     @Override
     public YandexPictureEntity getYandexPicture(long yandexPictureId) {
+        Log.d(TAG, "getYandexPicture: yandexPictureId = " + yandexPictureId);
         return YandexPictureEntity.findById(YandexPictureEntity.class, yandexPictureId);
     }
